pyproject.toml
[project]
name = "parkly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.14.*"
dependencies = [
    "loggerizer==1.0.0",
    "fastapi[standard]==0.115",
    "uvicorn[standard]==0.34",
    "python-ulid==3.0",
    "pydantic-settings==2.12",
]

[dependency-groups]
dev = ["debugpy==1.8", "pre-commit==4.1", "pytest==8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/parkly"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.14"
extraPaths = ["src"]

[tool.ruff]
line-length = 88
